Moved DataCollectorId and corresponding validator to its own folder in Concepts. Set inputvalidator for DataCollectorId
