+++ Source/UserManagement/Concepts/Concepts.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="Dolittle.Concepts" Version="2.0.0-alpha2*" />
+    <PackageReference Include="Dolittle.SDK.Validation" Version="2.0.0-alpha2.208" />
   </ItemGroup>
 
 </Project>
