+++ Source/UserManagement/Concepts/DataCollector/DataCollectorId.cs
@@ -1,7 +1,7 @@
 using System;
 using Dolittle.Concepts;
 
-namespace Concepts
+namespace Concepts.DataCollector
 {
     public class DataCollectorId : ConceptAs<Guid>
     {
