+++ Source/UserManagement/Concepts/DataCollector/DataCollectorIdValidator.cs
@@ -1,0 +1,14 @@
+using Dolittle.Validation;
+using FluentValidation;
+
+namespace Concepts.DataCollector
+{
+    public class DataCollectorIdValidator : InputValidator<DataCollectorId>
+    {
+        public DataCollectorIdValidator()
+        {
+            RuleFor(id => id.Value)
+                .NotEmpty().WithMessage("Data Collector Id cannot be empty");
+        }
+    }
+}
