+++ Source/UserManagement/Domain/DataCollector/Changing/ChangeBaseInformation.cs
@@ -1,5 +1,6 @@
 using System;
 using Concepts;
+using Concepts.DataCollector;
 using Dolittle.Commands;
 
 namespace Domain.DataCollector.Changing
