+++ Source/UserManagement/Domain/DataCollector/Changing/ChangeLocationValidator.cs
@@ -9,7 +9,8 @@ namespace Domain.DataCollector.Changing
         public ChangeLocationValidator()
         {
             RuleFor(_ => _.DataCollectorId)
-                .NotEmpty().WithMessage("Data Collector Id must be set");
+                .NotEmpty().WithMessage("Data Collector Id must be set")
+                .SetValidator(new DataCollectorIdValidator());
 
             RuleFor(_ => _.Location)
                 .Cascade(CascadeMode.StopOnFirstFailure)
