+++ Source/UserManagement/Domain/DataCollector/Changing/ChangePreferredLanguageValidator.cs
@@ -1,3 +1,4 @@
+using Concepts.DataCollector;
 using Dolittle.Commands.Validation;
 using FluentValidation;
 
@@ -7,6 +8,10 @@ namespace Domain.DataCollector.Changing
     {
         public ChangePreferredLanguageValidator()
         {
+            RuleFor(_ => _.DataCollectorId)
+                .NotEmpty().WithMessage("Data Collector Id must be set")
+                .SetValidator(new DataCollectorIdValidator());
+
             RuleFor(_ => _.PreferredLanguage)
                 .Cascade(CascadeMode.StopOnFirstFailure)
                 .IsInEnum().WithMessage("Preferred Language must be valid");
