+++ Source/UserManagement/Domain/DataCollector/PhoneNumber/AddPhoneNumberToDataCollector.cs
@@ -1,6 +1,4 @@
-using System;
-using System.Collections.Generic;
-using Concepts;
+using Concepts.DataCollector;
 using Dolittle.Commands;
 
 namespace Domain.DataCollector.PhoneNumber
