+++ Source/UserManagement/Domain/DataCollector/PhoneNumber/RemovePhoneNumberFromDataCollector.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Concepts;
+using Concepts.DataCollector;
 using Dolittle.Commands;
 
 namespace Domain.DataCollector.PhoneNumber
