+++ Source/UserManagement/Domain/DataCollector/PhoneNumber/RemovePhoneNumberFromDataCollectorValidator.cs
@@ -1,3 +1,4 @@
+using Concepts.DataCollector;
 using FluentValidation;
 
 namespace Domain.DataCollector.PhoneNumber
@@ -7,7 +8,8 @@ namespace Domain.DataCollector.PhoneNumber
         public RemovePhoneNumberFromDataCollectorValidator()
         {
             RuleFor(_ => _.DataCollectorId)
-                .NotEmpty().WithMessage("Data Collector Id is required");
+                .NotEmpty().WithMessage("Data Collector Id must be set")
+                .SetValidator(new DataCollectorIdValidator());
 
             RuleFor(_ => _.PhoneNumber)
                 .NotEmpty().WithMessage("Phone Number is required")
