+++ Source/UserManagement/Read/DataCollectors/DataCollector.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Concepts;
+using Concepts.DataCollector;
 using Dolittle.ReadModels;
