+++ Source/UserManagement/Read/DataCollectors/DataCollectors.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Concepts;
+using Concepts.DataCollector;
 using Infrastructure.Read.MongoDb;
 
 namespace Read.DataCollectors
