+++ Source/UserManagement/Read/DataCollectors/Queries/DataCollectorById.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using Concepts;
+using Concepts.DataCollector;
 using Dolittle.Queries;
 
 namespace Read.DataCollectors.Queries
