+++ Source/UserManagement/Web/Controllers/DataCollectorsController.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Concepts;
+using Concepts.DataCollector;
 using Dolittle.Queries;
 using Dolittle.Queries.Coordination;
 using Read.DataCollectors.Queries;
